+++ sample/src/main/kotlin/ca/allanwang/kau/sample/ImageActivity.kt
@@ -1,0 +1,8 @@
+package ca.allanwang.kau.sample
+
+import ca.allanwang.kau.imagepicker.ImagePickerActivityBase
+
+/**
+ * Created by Allan Wang on 2017-07-04.
+ */
+class ImageActivity : ImagePickerActivityBase()