+++ sample/src/main/kotlin/ca/allanwang/kau/sample/MainActivity.kt
@@ -212,7 +212,7 @@ class MainActivity : KPrefActivity() {
         when (item.itemId) {
             R.id.action_settings -> startActivity(AnimActivity::class.java, clearStack = true)
             R.id.action_email -> sendEmail(R.string.your_email, R.string.your_subject)
-            R.id.test -> prefHolder.showNext()
+            R.id.test -> startActivity(ImageActivity::class.java, transition = true)
             else -> return super.onOptionsItemSelected(item)
         }
         return true